Convert damage modifiers to integer percentages.

+++ OpenRA.Game/Traits/TraitsInterfaces.cs
@@ -155,7 +155,7 @@ namespace OpenRA.Traits
 	}
 
 	public interface IRenderModifier { IEnumerable<IRenderable> ModifyRender(Actor self, WorldRenderer wr, IEnumerable<IRenderable> r); }
-	public interface IDamageModifier { float GetDamageModifier(Actor attacker, DamageWarhead warhead); }
+	public interface IDamageModifier { int GetDamageModifier(Actor attacker, DamageWarhead warhead); }
 	public interface ISpeedModifier { decimal GetSpeedModifier(); }
 	public interface IFirepowerModifier { float GetFirepowerModifier(); }
 	public interface ILoadsPalettes { void LoadPalettes(WorldRenderer wr); }
